Repo cleanup | .gitignore & .gitattributes
Adds a .gitattributes file with a setting to enforce automatic
line ending normalization whether you're on Windows or Unix like OS.

As well as updating the .gitignore file with a more comprehensive
and controlled approach for working with config files. You may now work
with the git repo even in your game directory! Only the 'config' &
'scripts' folders will ever be included.

Every file not directly modified for the modpack has been excluded
from tracking (with the exception of HQM distributed outside github)

If you happen to need to modify any of the files mentioned in
.gitignore, feel free to do so and make sure to remove it from
.gitignore as well. Otherwise you'll not be able to stage the file
without `git add --force`.

Lastly, add files to .gitignore when a mod is added for best
effect.