Remove grain recipes
After updating, NEI still showed that rye + mortar and pestle creates a master carpentry hammer.  Since I can't seem to find these items in game anymore, I think they may be the source of this tomfoolery.  I've also updated the item ID for glass fittings (which don't appear to be in any recipes that I can find). 

I think this is the source of: