Remove recipes for deprecated extratrees items
All the recipes commented out were removed when extratrees removed the distillery and other booze making machines.  As a result the item ID for glass fittings was changed (this has also been updated).  

For whatever reason glass fittings are not used in any recipes that I can find.  May be something to consider when modifying recipes in the future.