Fix for Issue 2967
Description - I ended up fixing all of the ardite recipes. All of the recipes worked with the Tinkers Construct version of the Ardite Ingot but not the GregTech Ardite Ingot. I simply swapped this in the Tinkers-Construct.zs script. If you'd prefer these commented out and added to the GregTech.zs Extruder section let me know and I can do that instead. If you have any questions please let me know either on Issue 2967 or in Discord.