Fix for Issue 3790 Quest ID 938
Originally updated 'Damage' field from 26 to 27 but the quest still wouldn't turn in. This commit I removed the 'Damage' field all together which allows the electric treetap to be detected regardless of it's damage data. I think it works a little differently than most damage data in the sense that durability and charge seem to be interconnected somehow.