Update Biome-o-Plenty.zs
If you include a Bucket of Water and it also looks like Clay Bucket of Water you automatically get the bucket back without needing to script it in. This why it was duplicating. 
This not the case with all mods but it's the case here. You can also add a .noReturn() for those times you want it to be consumed.

This fixes:
#6424
#6354
#6256
#6230
#5727
#5591
#6089
#6514