Adds (almost) all End ores to meteor
Want almost all the ores the end has to offer has to offer? Well look no further!
Has a size of 18, costs 3.25M LP, and costs a HV Circuit Assembler. I can't add 'Sheldonite' as for some reason, the game doesn't want to properly register it, as a result, the other type of ore will be Platinum