Meteor-tastic changes (#10578)
* Adjusted Mysterious Crystal meteor (buff)

Changed name to MysteriousMeteor, adjusted item cost from a LuV Field Generator to an LuV Recycler, and changed the glass to Enceladus rock and buffed the MC value

* EV tiered lead-related meteor

Costs 1 EV assembling machine and 2M LP.
Yields a size 13 meteor of lead-related stuff in endstone variants.

* Nerf (cost) bat towards Super Plat Meteor

It now costs a IV scanner instead of a IV emitter

* Even bigger late-game "indium" meteor

Went through all the trouble to get the wetware supercomputer? Here's your wondrous reward.
Costs a wetware supercomputer (UV), 50M LP, and yields a 32 sized meteor full of lead, galena, silver, and a very trace amount of pure indium

* Nerf bat again to Plat meteor

Costs an IV mass fab now instead of a IV scanner (first original cost was an IV sensor)

* Nerfed end-game meteor a bit

Adjusted various values of the meteor based on feedback of a friend. This will indirectly buff & nerf some other values

* Fixed whoopies

* Re-adjusted super plat meteor viability

Made it a radius of 14 (it's still really op) but made the cost 12.5M instead of 25M. It's now more realistically possible to obtain with a T5 altar & orb, or rather, you can kill some lesser demons and obtain the T5 orb and a T5 altar (which requires 4 beacons)

* New pure naquadah meteor

It is, however, rather tedious to obtain and is meant as a treat for those who will go this far to get it.
Requires 25M LP and a Crystal Processor Mainframe (UV) circuit. Primarily great for Tiberium

* Fixed another oversight in meteor

Heh, whoopies

* Plat meteor costs an IV scanner for real

It instead costed a IV mass fabricator

* Added radioactively pure meteor

Contains all but naquadah GT elements! Costs an IV Field Generator and 2.5M LP

* Fixed minor error

It's 'uraninite,' not 'uranite,' heh

* Indirectly made plutonium more rarer in meteor

And adjusted some other various stats

* Radio. meteor cost IV mass fab but increased sized

It now costs an IV mass fab since the field generator was already used in another meteor

* Readjusted stats of IndiumGreaterVariant

Shifted it around to be less insane, and yield potentially more than 1 indium ore (lmao) out of a radius of 18

* Nerfed radius of ConcentratedLeadEV meteor

This is a rather large nerf, but I realized it's going to be really insane

* Europium & diffusion to end-game meteor

This also indirectly nerfs the rest of the stats of the meteor itself with diffusion. Also adjusted some of the other stats like the radius

* CallistoMeteor mega radius buff

Is now radius 20 instead of 16

* Mega nerfed the infamously giant meteor

Now has a radius of 22 instead of 28. Also re-adjusted some various stats around

* Buffed OW meteor V.2 radius

Now is a radius of 18 instead of 16