Add shapeless recipe for Railcraft tank blocks
The shapeless recipes added here (which have been added for all the main tank blocks) are designed to address the problem of Railcraft tank blocks, when picked up, being unable to be merged back into a stack of blocks that have yet to be placed in the world. This is because of NBT data being retained when a block is broken. The shapeless recipes reset the NBT data, allowing blocks to become stackable once again.

Tested in a 2.1.3.4-25 environment.