more lp changes (#11643)
* more lp changes

* change ItemDetectorCover to GT ItemFilter in passive supplier

ItemDetectorCover is an item with way too many steps to craft and is used way too rare. Even if ItemFilter is a bit more expensive, it is overall an relieving change. The batch per craft is increased as a compensate for price.

Signed-off-by: Glease <4586901+Glease@users.noreply.github.com>

* fix typo in c3958a6b

Signed-off-by: Glease <4586901+Glease@users.noreply.github.com>

* firewall isn't a reactor vessel and doesn't need to be rad proof

Signed-off-by: Glease <4586901+Glease@users.noreply.github.com>

* Buff late game recipes for chassis

to spend 1 EV circuit is a bit too expensive and plays poorly with human mind. buff it to encourage people to use this at all.

Signed-off-by: Glease <4586901+Glease@users.noreply.github.com>

* remove nether star plate

seriously, why tf does a mk5 need nether stars

Signed-off-by: Glease <4586901+Glease@users.noreply.github.com>

* reduce price of crafting module

the project red counterpart is 1 MV circuit per chip. No reason to make the LP one so much more expensive

Signed-off-by: Glease <4586901+Glease@users.noreply.github.com>

* maybe too cheap

Signed-off-by: Glease <4586901+Glease@users.noreply.github.com>

* i take it back. i forget these compoents also use circuits

Signed-off-by: Glease <4586901+Glease@users.noreply.github.com>

* Fix Request Pipe Not Being MV

* make crafting modules based off empty modules instead of ItemSink

Signed-off-by: Glease <4586901+Glease@users.noreply.github.com>

Signed-off-by: Glease <4586901+Glease@users.noreply.github.com>
Co-authored-by: Steelux8 <steelux7@gmail.com>