Quest: ETOOMANYKEYS: add a note about the waypoint manager
This shortcut is not mentioned in any other quest or controls, but it
saves a ton of time and nerves if you use teleportation because it
allows you to instantly open Waypoints rather than first open the
JourneyMap and then manually click on the Waypoints button.

Aside from that, fix typos and improve wording.