ANewAlloyMagnali: update
The part about getting Magnesium mentioned a mineralsand vein that
requires MV electrolyzer to get Magnesium out of it. But it's an
LV quest, so it's irrelevant - there are other veins and resources
at MV that can give Magnesium, so it was replaced with a sentence
about NEI.

The quest also mentioned Gypsum as a resource for BBF, but BBF is
long done at this point, so it was deleted.

Also reword the quest TimetoGetSomeSte.