Another Round of Meteor Shenanigans
This will contain even more changes, adjustments, and additions to newer meteors. This new meteor is just a start and there'll be much more content down the line. This is not meant to be merged into master without careful balancing discussions :v

For the balance of the Raw Tengam meteor, the ratios are 580 salt ore & 1 tengam respectively, this should, in theory, make the overall ratio of the ore 4.20% average (unless I did my math badly, heh).