Tengam made even rarer
MAJORLY adjusted the stats and also costs. This definitely needs to be tested in-game to be made sure that it actually spawns in. However, what this'll create is a meteor with a 0.13% Tengam ratio (~23 with a size of 16). If you use Orbis Terrae, the amount of ores increases to ~33.

Even if you use an ore driller and Orbis Terrae to minmax the amount of ore, this is unholy expensive & horrible to do that this is only realistically done in meme runs