Master tier meteor 1
Features a very expensive item but absurd resource gain. This will not yield Dilithium to prevent cheesing the Space Elevator. Yields almost every considered gem in the game, with such an enormous yield that you probably only want to do this once, though it's probably never going to see the light of day due to void miners. But hey, fun to complete, eh?