Reduced size of meteor to prevent mega lag spike
There were about 463,268 blocks in a size 48 meteor. All of it was GT ores. On a public server, this is the equivalent of loading up the old Deep Dark but somehow there were almost half a million GT ores in render distance

The size is now 32 instead of 48
Also, LP is 1 since you probably wouldn't even touch this meteor until post-Armok blood orb. Unless you're a dare devil!