Create Trigger Quest: Nether Skip (#17244)
* Create Trigger Quest: Nether Skip

Note: This was written in 2.6.1's questbook, and I have not written a quest before. I tried to match how other quests were made in the discord, so there might need to be edits to fix this.


This adds a Trigger Quest to skip "THE NETHER" & "How To Make Rubber" by obtaining 128 Rubber Bars in your inventory. Doing so will unlock "Wire-Wrap".

This is so players who decided to progress through the steam age without going to the nether (fear of nether mobs, wanting to build renewable methods early through XP Buckets, etc.) can still progress through the questbook.

* Update TriggerNetherSki-qFco17HGS0SxmpVn-uWJZQ==.json

---------

Co-authored-by: Martin Robertz <dream-master@gmx.net>