Update InGameInfo.xml - Adding center chunk icon & removing unecessaey chunk checks (#17729)
* Update InGameInfo.xml - Adding center chunk icon & removing unecessary chunk checks

Noticed that the current .xml file has an unnecessary chunk coordinate check (chunk coordinates, not relative coordinates within a chunk) for the "center of the chunk" line. I removed that part and added in the obsidian tough binding icon to be placed next to this line.

I would've added some sort of red tough binding icon but I could not figure out the proper file path for like an ignis tough binding. So obsidian it is.

* fix weird icon tab size

* Update InGameInfo.xml - fix tab/space formatting