Tree Breeding QB Changes
Based on the suggestion in ticket #19892 , I've updated the Saplings in the Forestry Tab to reflect how many you actually need to grow the trees at the end of the line.

This should give the player a proper idea of how many saplings they need to make if they want to grow the tree that they are aiming for.