Split the Anvil Quest into 2
Split the Anvil quest into 2, so you don't have to craft 2 at one time to finish the quest. You can craft one, and then a second quest unlocks to either craft a second one or just retrieve it. This then doesn't force the player to have to craft an extra anvil when they only need for the Forge Hammer.

Closes: #19473