Added a quest for the Redstone Sniffer
Added a quest for the Redstone Sniffer. Includes information about what the sniffer does.

I put it into EV since that's when you can craft it, and this is around the time when you would start to use covers for power, which this tool will then be a helpful tool for you.