Solar Factory Quest
Added a quest for the Solar Factory on the Multiblock Goals Tab. I linked it to the HV Assembler quest since you require one to actually make the multi itself.

I didn't add a quest reward because I'm not sure where this multi is really "placed." Since you won't make it till like ZPM, you shouldn't get a ZPM lootbag if its a HV unlockable multi. This could be changed though if need be.